+++ GrowIT/Core/Network/Auth/Token/AuthPlugin.swift
@@ -210,7 +210,7 @@ final class AuthPlugin: PluginType {
             print("🗑️ Kingfisher 디스크 캐시 초기화 완료")
         }
 
-        let loginVC = LoginViewController()
+        let loginVC = MainLoginViewController()
         let nav = UINavigationController(rootViewController: loginVC)
         if let scene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
            let window = scene.windows.first {
