--- GrowIT/Feat/Login/Presentation/Model/LoginModel.swift
@@ -1,13 +1,0 @@
-//
-//  LoginModel.swift
-//  GrowIT
-//
-//  Created by 강희정 on 1/29/25.
-//
-
-import Foundation
-
-struct LoginModel {
-    let email: String
-    let password: String
-}
