+++ GrowIT/Feat/Login/Presentation/ViewController/MainLoginViewController.swift
@@ -10,7 +10,7 @@ import Foundation
 import SnapKit
 import AuthenticationServices
 
-class LoginViewController: UIViewController {
+class MainLoginViewController: UIViewController {
     // MARK: - Properties
     let authService = AuthService()
     private lazy var kakaoLoginHelper = KakaoLoginHelper()
@@ -243,14 +243,14 @@ class LoginViewController: UIViewController {
     
 }
 
-extension LoginViewController: ASAuthorizationControllerPresentationContextProviding {
+extension MainLoginViewController: ASAuthorizationControllerPresentationContextProviding {
     // 인증창을 보여주기 위한 메서드 (인증창을 보여 줄 화면을 설정)
     func presentationAnchor(for controller: ASAuthorizationController) -> ASPresentationAnchor {
         self.view.window ?? UIWindow()
     }
 }
 
-extension LoginViewController: ASAuthorizationControllerDelegate {
+extension MainLoginViewController: ASAuthorizationControllerDelegate {
     // 로그인 실패 시
     func authorizationController(controller: ASAuthorizationController, didCompleteWithError error: any Error) {
         print("로그인 실패", error.localizedDescription)
