+++ GrowIT/Feat/Login/Signup/ViewControllers/errorVC/ChangePasswordErrorViewController.swift
@@ -1,5 +1,5 @@
 //
-//  EmailVerificationErrorViewController.swift
+//  ChangePasswordErrorViewController.swift
 //  GrowIT
 //
 //  Created by 강희정 on 1/26/25.
@@ -7,15 +7,15 @@
 
 import UIKit
 
-class EmailVerificationErrorViewController: UIViewController {
+class ChangePasswordErrorViewController: UIViewController {
     
     let errorView = ErrorView().then {
         $0.configure(
             icon: "trashicon",
-            fisrtLabel: "회원가입을 그만할까요?",
-            secondLabel: "해당 화면을 나가면 처음부터 다시 진행합니다\n그래도 화면을 나갈까요?",
+            fisrtLabel: "비밀번호 재설정을 그만할까요?",
+            secondLabel: "해당 화면을 나가면 이메일 인증을 다시 진행합니다\n그래도 화면을 나갈까요?",
             firstColor: .gray900,
-            secondColor: .gray700,
+            secondColor: .gray600,
             title1: "나가기",
             title1Color1: .gray400,
             title1Background: .gray100,
@@ -26,6 +26,8 @@ class EmailVerificationErrorViewController: UIViewController {
             viewColor: .white
         )
     }
+
+    
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -56,10 +58,10 @@ class EmailVerificationErrorViewController: UIViewController {
         if let navigationController = self.presentingViewController as? UINavigationController {
             dismiss(animated: true) {
                 // LoginViewController로 이동
-                let emailLoginVC = EmailLoginViewController()
-                let loginVC = LoginViewController()
-                navigationController.setViewControllers([loginVC, emailLoginVC], animated: false)
+                let loginVC = MainLoginViewController()
+                navigationController.setViewControllers([loginVC], animated: true)
             }
         }
     }
+    
 }
